chore: move more into pagination component
